requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.1
django-allauth==0.50.0
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
filelock==3.4.2
idna==3.3
oauthlib==3.2.0
Pillow==9.0.1
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.13.0
whitenoise==6.0.0
